Generalize report event handling
Introduce a Reporter interface that can be implemented in multiple ways.

Provide two Reporter implementations:

* one that logs JSON to a writer, as the monolithic JSONMiddleware did
* one that filters events based on a provided function, and calls
  another reporter selectively

Move the HTTP handling into a separate middleware that reports events to
any Reporter. Re-implement the JSONMiddleware using it.

Remove the selective JSON middleware again, it is now a one-liner that
can be constructed as needed. Do that in the tests to demonstrate how it
all fits together.

Signed-off-by: Matthias Rampke <matthias@rampke.de>